pyproject.toml
[project]
name = "spotify-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "google-auth>=2.48.0",
    "google-cloud-storage>=3.9.0",
    "grpcio>=1.78.0",
    "httpx==0.25.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn==0.24.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
